refactor(main): split main.py into focused modules; audit fixes (#2738)
main.py (1514 lines) is now a thin composition root; its sub-concerns
moved to lifecycle.py (Lifecycle, lifespan, setup_logfire),
middleware.py (LiveCORSMiddleware, InFlightRequests/Middleware),
static.py (StaticFiles mounts + no-cache middleware), and
bind_safety.py (no-auth loopback-bind gate). All former klangk.main
names are re-exported for existing callers/tests.

Inline audit fixes:
- Register the no-cache middleware once in build_app instead of inside
  setup_static_files: Starlette forbids add_middleware after the app
  starts serving, so the SIGHUP frontend remount crashed the recycle
  on a live server.
- _remount_frontend also drops the stale /branding mount so a reload
  re-resolves it from the new settings.
- The password-mode admin lockout guard and agent-handle collision
  guard raise ConfigurationError (not bare RuntimeError) so the
  launcher exits EX_CONFIG instead of restart-looping (#2666).
- Drop leading underscores from moved helpers
  (_is_loopback_bind, _agent_principal_error_handler) per naming rule.

Backend suite: 5535 passed, coverage 100% (klangk.main 700->65 stmts).
